requirements.txt
earthengine-api>=0.1.300
pyyaml>=6.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
